requirements.txt
tweepy==4.14.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
jieba==0.42.1
textblob==0.17.1
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
transformers==4.35.2
torch==2.1.0
langdetect==1.0.9
pytz==2023.3
openpyxl==3.1.2
jsonlines==4.0.0